pyproject.toml
[tool.poetry]
name = "samplepoetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-binance = "^1.0.19"
binance-futures-connector = "^4.0.0"
coverage = "^7.4.4"
numpy = "^1.26.4"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
